+++ backend/src/domain/models/document.py
@@ -1,6 +1,6 @@
-from dataclasses import dataclass
-from datetime import datetime
 from typing import Optional
+from datetime import datetime
+from dataclasses import dataclass
 
 @dataclass
 class Document:
