+++ backend/src/domain/models/vehicle.py
@@ -1,5 +1,5 @@
-from dataclasses import dataclass
 from datetime import datetime
+from dataclasses import dataclass
 
 @dataclass
 class Vehicle:
