--- backend/src/infrastructure/web/configure_cors.py
@@ -1,6 +1,0 @@
-from flask_cors import CORS
-
-
-def configure_cors(app):
-    CORS(app, supports_credentials=True)
-    return app
